Refactored helper modulue, created new threader modulue to handle thread logging, fixed incorrect headed_support flag
